refactor: Clearer seperation of components vs UI componenets
