+++ frontend/src/lib/components/booru-search.svelte
