+++ frontend/src/lib/components/json-code-block.svelte
